Create Painting a Wall
"It is known that a can of paint has a cost of C and is capable of painting an area of ​​M square meters. Write a program that reads the width L, the height A of a wall, the C value of a can of paint and the yield M of this can. After that, print how many paint cans are needed and the total cost (with two decimal places)."