Create Chocolate
"Osmar loves chocolates and goes to the store with N money in his pocket. The price of each chocolate is C. The store offers a discount: for every M (always greater than 1) packages he gives to the store, he gets a free chocolate. How many chocolates can Osmar eat? For example:

For N=10, C=2, M=5, he can buy 5 chocolates for $10 and exchange the 5 packages for 1 more chocolate, making the total number of chocolates he can eat 6.
Make a program that reads N, C and M and prints the amount of chocolates that Osmar can eat."