Create Ana's Number
"Ana is very fond of mathematics and playing with numbers. She defined the concept of "Anna's Number. If a non-negative integer N is the product of three consecutive integers then N is a "Anna's Number".

Write a program that reads a non-negative integer N and prints "S" if it's "Anna's Number" and "N" otherwise. For example, 120 is an "Ana's Number" as it is the result of multiplying 4 x 5 x 6."