Create Million Circle
"There are N people sitting in a circle. People are numbered in order, from 1 to N. In the center of the circle is a billionaire who likes to distribute millions around. He will choose one person to win a million. For this he will eliminate from the circle N - 1 people, following two rules:
The billionaire eliminates one person from the circle and always skips the next one, until only one person remains, who is the winner of the prize.
The billionaire always starts by eliminating person number 2.

For example: Considering N = 1000, the billionaire eliminates person number 2, followed by 4, 6, 8, ...., 996, 998, 1000. At this point, he skips number 1 and eliminates person number 3, number 7 and so on.

Given this scenario, make a program that, given the number N of people in the circle, the program informs:
What is the number of the person who will win the prize; It is
What is the xth person to be eliminated?"