Create Bowling Points System
"Bowling is a sport practiced with a heavy ball and aims to throw the ball down a track, knock down 10 pins on the opposite side of the track arranged in a triangular formation (https://www.infoescola.com/esportes/boliche/).
The formula for counting points in bowling has the following variables (https://boliche.com.br/esporte-boliche/contagem-dos-pontos-no-boliche/):
The points are the sum of the knocked down pins.
Except when striking (knocking down 10 pins on the 1st ball) or spare (knocking down 10 pins on the two balls played)
If you strike, you get a bonus on the 2 balls played next. If you make a spare, you get a bonus on the next ball played. The bonus is equal to the number of pins knocked down.
The score of a match can go from zero (when no pin is knocked down in the ten moves or "frames") up to the maximum possible of 300 points, that is, 12 consecutive "strikes". Supposedly, as each game has 10 “frames” (moves), only 10 “strikes” would be possible. However, if the player knocks down all the pins in the first shot of the 10th "frame", he has the right to play two more balls, being able to complete 12 "strikes" in the same line.

Write a program that reads the number of pins dropped by a bowler in each round and prints:
the sequence of knocked down pins (according to the input and output examples and the points counting notes - https://boliche.com.br/esporte-boliche/contagem-dos-pontos-no-boliche/);
the player's final score."